move functions to helpers function
