+++ utils/helpers.py
@@ -1,0 +1,40 @@
+import torch.nn as nn
+from torch.utils.data import DataLoader
+
+
+def parallel_model(model):
+    return nn.DataParallel(model)
+
+
+def load_missing(model, pretrained_dict):
+    model_dict = model.state_dict()
+    pretrained_dict = {k: v for k, v in pretrained_dict.items() if
+                       k in model_dict.keys() and v.size() == model_dict[k].size()}
+    missed_params = [k for k, v in model_dict.items() if not k in pretrained_dict.keys()]
+
+    print('loaded params/tot params:{}/{}'.format(len(pretrained_dict), len(model_dict)))
+    print('miss matched params:', missed_params)
+    model_dict.update(pretrained_dict)
+    model.load_state_dict(model_dict)
+    return model
+
+
+def show_lr(optimizer):
+    return ','.join(['{:.6f}'.format(param_group['lr']) for param_group in optimizer.param_groups])
+
+
+def dataset2dataloader(dataset, batch_size, num_workers, shuffle=True):
+    loader = DataLoader(dataset,
+                        batch_size=batch_size,
+                        num_workers=num_workers,
+                        shuffle=shuffle,
+                        drop_last=False,
+                        pin_memory=True)
+    return loader
+
+
+def add_msg(msg, k, v):
+    if msg:
+        msg += ','
+    msg += k.format(v)
+    return msg
